Some ability fixes (#969)
* bramble shell iframe bypass
* description fix in split arrow
* Removed creeper's delights from depths loot
* Steel Stallion awarding strike fix
* new depths room

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/depths/abilities/steelsage/DepthsSplitArrow.java
@@ -76,7 +76,7 @@ public class DepthsSplitArrow extends DepthsAbility {
 
 	@Override
 	public String getDescription(int rarity) {
-		return "When you shoot an enemy with an arrow, the nearest enemy within " + SPLIT_ARROW_CHAIN_RANGE + " blocks takes " + DepthsUtils.getRarityColor(rarity) + DepthsUtils.roundPercent(DAMAGE_MOD[rarity - 1]) + "%" + ChatColor.WHITE + " of the arrow's damage. Bypasses iframes.";
+		return "When you shoot an enemy with an arrow, the nearest enemy within " + SPLIT_ARROW_CHAIN_RANGE + " blocks takes " + DepthsUtils.getRarityColor(rarity) + (int) DepthsUtils.roundPercent(DAMAGE_MOD[rarity - 1]) + "%" + ChatColor.WHITE + " of the arrow's damage. Bypasses iframes.";
 	}
 
 	@Override
